Remove the storage dependency from the Timeline component
In the next version of T-Inbox, the tweets will no longer be stored
in the localstorage.

This approach does not scale, since the list of tweets is never
flushed from the Chrome storage.

Also, if the user has the extension installed in two different
machines the list of unarchived tweets is not persisted between
both.

The new version of T-Inbox will rely on the server for archiving
and fetching the unarchived list of tweets