Refactor the background process to rely on the API for counting tweets
The API endpoint that returns the tweets has now a 'total' attribute in the payload, this attribute is used to display in the extension badge how many tweets the user has to read.

In this commit we are also allowing the extension and the web app to send messages to the background process, so it can update the extension badge based on the user interactions. For example, if the user archives a tweet, the extension is going to emmit a 'decreaseTweetCount' message, in the other side the background process intercepts this message and handle it updating the extension badge.

We are also refactoring the storage module, removing all the deprecated methods.